future: drop internal call_then_impl
call_then_impl was used as a internal customization point for supporting
API level 3 and below, where when_all_succeed returned a variadic
future. Since that's no longer supported, the customization point
is removed and its contents is inlined into its callers.

It also served a minor role in the concept definition for
future::then(), so that's adjusted too.

Closes #1575